requirements.txt
django
honcho
djangorestframework
psycopg2
django-webpack-loader
